o Fixes some compiler warnings

+++ MapView/Map/RMQuadTree.m
@@ -529,7 +529,7 @@
 
                 _cachedClusterAnnotation = [[RMAnnotation alloc] initWithMapView:_mapView
                                                                      coordinate:clusterMarkerCoordinate
-                                                                       andTitle:[NSString stringWithFormat:@"%d", enclosedAnnotationsCount]];
+                                                                       andTitle:[NSString stringWithFormat:@"%lu", (unsigned long)enclosedAnnotationsCount]];
                 _cachedClusterAnnotation.annotationType = kRMClusterAnnotationTypeName;
                 _cachedClusterAnnotation.userInfo = self;
